refactor: Move Media fetchers into Media file
