+++ src/lib/omerlo/reader/endpoints/profileType.ts
@@ -1,14 +1,7 @@
-import { parseMany, type ApiAssocs, type ApiData, type PagingParams } from '$reader/utils/api';
+import { type ApiAssocs, type ApiData } from '$reader/utils/api';
 import type { LocalesMetadata } from '$reader/utils/response';
-import { requestPublisher } from '$reader/utils/request';
 import { buildMeta } from '$reader/utils/parseHelpers';
 
-export const profileTypeFetchers = (f: typeof fetch) => {
-  return {
-    getProfileType: getProfileType(f),
-    listProfileTypes: listProfileTypes(f)
-  };
-};
 
 export interface ProfileTypeSummary {
   id: string;
@@ -60,19 +53,5 @@ export function parseProfileType(data: ApiData, assocs: ApiAssocs): ProfileType 
     hasState: data.has_state || false,
     hasCity: data.has_city || false,
     hasStreet: data.has_street || false
-  };
-}
-
-export function getProfileType(f: typeof fetch) {
-  return async (id: string) => {
-    const opts = { parser: parseProfileType };
-    return requestPublisher(f, `media/profile-types/${id}`, opts);
-  };
-}
-
-export function listProfileTypes(f: typeof fetch) {
-  return async (params?: Partial<PagingParams>) => {
-    const opts = { parser: parseMany(parseProfileTypeSummary), queryParams: params };
-    return requestPublisher(f, `media/profile-types`, opts);
   };
 }
