+++ src/lib/omerlo/reader/endpoints/projects.ts
@@ -8,18 +8,10 @@ import {
   parseProfileDescription,
   type ProfileDescription
 } from './profiles';
-import { type ApiData, parseMany, type PagingParams, type ApiAssocs } from '$reader/utils/api';
+import { type ApiData, type ApiAssocs } from '$reader/utils/api';
 import { getAssoc, getAssocs } from '$reader/utils/assocs';
-import { requestPublisher } from '$reader/utils/request';
 import { buildMeta } from '$reader/utils/parseHelpers';
 import type { ProfileType } from './profileType';
-
-export const projectFetchers = (f: typeof fetch) => {
-  return {
-    listProjects: listProjects(f),
-    getProject: getProject(f)
-  };
-};
 
 export interface ProjectSummary {
   id: string;
@@ -75,19 +67,5 @@ export function parseProject(data: ApiData, assocs: ApiAssocs): Project {
     contact,
     address,
     description
-  };
-}
-
-export function getProject(f: typeof fetch) {
-  return async (id: string) => {
-    const opts = { parser: parseProject };
-    return requestPublisher(f, `media/projects/${id}`, opts);
-  };
-}
-
-export function listProjects(f: typeof fetch) {
-  return async (params?: Partial<PagingParams>) => {
-    const opts = { parser: parseMany(parseProjectSummary), queryParams: params };
-    return requestPublisher(f, `media/projects`, opts);
   };
 }
